requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
fastapi==0.116.1
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
gspread==6.2.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
mysql-connector==2.2.9
numpy==2.2.6
oauthlib==3.3.1
openai==1.99.9
pandas==2.3.1
portalocker==3.2.0
proto-plus==1.26.1
protobuf==6.31.1
py-dotenv==0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydentic==0.0.1.dev3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-stdnum==2.1
pytz==2025.2
qdrant-client==1.15.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
